// types/LandingPageType.tsx
export interface LandingPageStatisticData {
    id: number;
    title: string;
    count: number;
    image: string;
    order: number;
    createdAt: string;
    updatedAt: string;
}

export interface LandingPageWithUsData {
    id: number;
    title: string;
    subtitle: string;
    image: string;
    order: number;
    createdAt: string;
    updatedAt: string;
}